Drop multiversion support in Python model wrapper
