+++ CMakeLists.txt
@@ -276,7 +276,7 @@ install(
 #
 set(MODEL_SCHEMA_PATH "${CMAKE_BINARY_DIR}/model-schema.yml")
 set(MODEL_JSONSCHEMA_PATH "${CMAKE_BINARY_DIR}/model-jsonschema.yml")
-set(PYTHON_GENERATED_MODEL_PATH revng/model/v1/_generated.py)
+set(PYTHON_GENERATED_MODEL_PATH revng/model/_generated.py)
 
 #
 # Enable CTest
