+++ python/CMakeLists.txt
@@ -113,7 +113,7 @@ python_module(TARGET_NAME python-top-internal-module WHEEL revng_internal
 # Install revng.model (including autogenerated classes)
 #
 set(PYTHON_MODEL_FILES revng/model/__init__.py revng/model/metaaddress.py
-                       revng/model/v1/__init__.py revng/model/v1/external.py)
+                       revng/model/__init__.py revng/model/external.py)
 python_module(
   TARGET_NAME
   python-model
