+++ python/revng/model/external.py
@@ -2,4 +2,4 @@
 # This file is distributed under the MIT License. See LICENSE.md for details.
 #
 
-propertyA: somethingsomething
+from .metaaddress import MetaAddress  # noqa: F401
