--- python/revng/model/v1/external.py
@@ -1,5 +1,0 @@
-#
-# This file is distributed under the MIT License. See LICENSE.md for details.
-#
-
-from ..metaaddress import MetaAddress  # noqa: F401
