+++ tests/CMakeLists.txt
@@ -4,5 +4,4 @@
 
 add_subdirectory(abi)
 add_subdirectory(pipeline)
-add_subdirectory(tuple-tree-generator)
 add_subdirectory(unit)
