--- tests/tuple-tree-generator/CMakeLists.txt
@@ -1,5 +1,0 @@
-#
-# This file is distributed under the MIT License. See LICENSE.md for details.
-#
-
-add_subdirectory(python)
