--- tests/tuple-tree-generator/python/v1_schema.yml
@@ -1,11 +1,0 @@
-#
-# This file is distributed under the MIT License. See LICENSE.md for details.
-#
-
-version: 1
-definitions:
-  - name: RootType
-    type: struct
-    fields:
-      - name: propertyA
-        type: string
