--- tests/tuple-tree-generator/python/v2.yml
@@ -1,5 +1,0 @@
-#
-# This file is distributed under the MIT License. See LICENSE.md for details.
-#
-
-propertyB: somethingsomething
